pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
pygame = "^2.1.2"
dearpygui = "^1.6.2"
numpy = "^1.24.2"
PyYAML = "^6.0"
rich = "^13.3.1"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"
python-lsp-server = {extras = ["yapf", "rope", "pyflakes"], version = "^1.5.0"}
toml = "^0.10.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"